+++ packages/components/src/components/LightBox/LightBox.tsx
@@ -47,6 +47,8 @@ export const LightBox = flowComponent("LightBox", (props) => {
 
   const propsContext: PropsContext = {
     ActionGroup: {
+      preserveOrder: true,
+      spacing: "m",
       className: styles.actionGroup,
       Button: { variant: "solid", color: "light-static" },
       tunnel: {