+++ packages/components/src/components/List/components/Items/components/Item/Item.tsx
@@ -1,7 +1,8 @@
 import { useGridItemProps } from "@/components/List/components/Items/components/Item/hooks/useGridItemProps";
 import { useList } from "@/components/List/hooks/useList";
 import ItemsGridListItemView from "@/views/ItemsGridListItemView";
-import type { FC, PropsWithChildren } from "react";
+import { mergeRefs } from "@react-aria/utils";
+import type { FC, PropsWithChildren, Ref } from "react";
 import { Suspense } from "react";
 import type { Key } from "react-aria-components";
 import styles from "./Item.module.scss";
@@ -10,10 +11,11 @@ import { ListItemSkeletonView } from "./components/ListItemSkeletonView/ListItem
 interface Props extends PropsWithChildren {
   id: Key;
   data: never;
+  triggerRef?: Ref<HTMLDivElement>;
 }
 
 export const Item = (props: Props) => {
-  const { id, data } = props;
+  const { id, data, triggerRef } = props;
   const list = useList();
 
   const itemView = list.itemView;
@@ -28,6 +30,11 @@ export const Item = (props: Props) => {
   const href = itemView.href ? itemView.href(data) : undefined;
   const hasAction = !!gridItemProps.onAction || !!href;
 
+  const existingRef = "ref" in gridItemProps ? gridItemProps.ref : undefined;
+  const gridItemPropsWithRef = triggerRef
+    ? { ...gridItemProps, ref: mergeRefs(existingRef, triggerRef) }
+    : gridItemProps;
+
   return (
     <ItemsGridListItemView
       id={id}
@@ -36,7 +43,7 @@ export const Item = (props: Props) => {
       target={itemView.target}
       hasAction={hasAction}
       isTile={list.viewMode.isTiles}
-      {...gridItemProps}
+      {...gridItemPropsWithRef}
     >
       <Suspense
         fallback={<ListItemSkeletonView viewMode={list.viewMode.value} />}