+++ packages/remote-react-components/CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.923](https://github.com/mittwald/flow/compare/0.2.0-alpha.922...0.2.0-alpha.923) (2026-07-16)
+
+### Features
+
+* **ActionGroup:** add preserveOrder prop for DOM-order layout ([#2652](https://github.com/mittwald/flow/issues/2652)) ([a0f14f1](https://github.com/mittwald/flow/commit/a0f14f11fb0a305e3589fa392f7d5dceacb70b00))
+
+# [0.2.0-alpha.922](https://github.com/mittwald/flow/compare/0.2.0-alpha.921...0.2.0-alpha.922) (2026-07-16)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.921](https://github.com/mittwald/flow/compare/0.2.0-alpha.920...0.2.0-alpha.921) (2026-07-16)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.920](https://github.com/mittwald/flow/compare/0.2.0-alpha.919...0.2.0-alpha.920) (2026-07-16)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.919](https://github.com/mittwald/flow/compare/0.2.0-alpha.918...0.2.0-alpha.919) (2026-07-16)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.918](https://github.com/mittwald/flow/compare/0.2.0-alpha.917...0.2.0-alpha.918) (2026-07-16)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.917](https://github.com/mittwald/flow/compare/0.2.0-alpha.916...0.2.0-alpha.917) (2026-07-16)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.916](https://github.com/mittwald/flow/compare/0.2.0-alpha.915...0.2.0-alpha.916) (2026-07-15)
+
+### Bug Fixes
+
+* **Docs:** remove stretch prop ([#2647](https://github.com/mittwald/flow/issues/2647)) ([2d527f9](https://github.com/mittwald/flow/commit/2d527f9a61dc444f7b70587aac764abff9ce8843))
+
+# [0.2.0-alpha.915](https://github.com/mittwald/flow/compare/0.2.0-alpha.914...0.2.0-alpha.915) (2026-07-15)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
+# [0.2.0-alpha.914](https://github.com/mittwald/flow/compare/0.2.0-alpha.913...0.2.0-alpha.914) (2026-07-15)
+
+### Features
+
+* **remote:** surface "No component found" error on the remote side ([#2651](https://github.com/mittwald/flow/issues/2651)) ([715fd75](https://github.com/mittwald/flow/commit/715fd7537cc776f8376ffc4634021071ba5fde18)), closes [#2441](https://github.com/mittwald/flow/issues/2441)
+
+# [0.2.0-alpha.913](https://github.com/mittwald/flow/compare/0.2.0-alpha.912...0.2.0-alpha.913) (2026-07-15)
+
+**Note:** Version bump only for package @mittwald/flow-remote-react-components
+
 # [0.2.0-alpha.912](https://github.com/mittwald/flow/compare/0.2.0-alpha.911...0.2.0-alpha.912) (2026-07-15)
 
 **Note:** Version bump only for package @mittwald/flow-remote-react-components